PrefMgr // Split its extension to a standalone file.
